+++ changelog.md
@@ -6,12 +6,15 @@
 
 - [#4790](https://github.com/ignite/cli/pull/4790) Remove global vars and struct placeholders.
 
+### Changes
+
+- [#4794](https://github.com/ignite/cli/pull/4794) Update `cosmossdk.io/log` version to `v1.6.1` and remove `bytedance/sonic` version override from template.
+
 ## [`v29.3.1`](https://github.com/ignite/cli/releases/tag/v29.3.1)
 
 ### Fixes
 
 - [#4793](https://github.com/ignite/cli/pull/4793) Use latest `bytedance/sonic` version to support Go 1.25.
-
 
 ## [`v29.3.0`](https://github.com/ignite/cli/releases/tag/v29.3.0)